feat(pdf-cropper): let user upload cropper data zip file if already present so that it can be used to pre generate images
File uploader of pdf cropper will can now accept pdf or zip files.
if it is zip file then it directly shows generate output dialog with the data of zip file, allowing user to download the pre generated version of zip instead.

if it is pdf then does the pdf cropper's indented job