fix(cbt-interface): fix current question timeSpent in testLogs of test interface
In the test logs, it was showing the the question's timespent value that had been set when question was changed, this limited to every action in current question to show the same timeSpent in test logs.
Now it will calculate the timeSpent on current question on that log instant to show the real time timeSpent in test logs.